Footer updated
