+++ package.json
@@ -11,11 +11,11 @@
     "tsc": "bash -c tsc --noEmit",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "lint:staged": "eslint"
-    
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.13.6",
     "@mui/styled-engine-sc": "^5.12.0",
@@ -23,6 +23,7 @@
     "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.10.1",
     "react-router-dom": "^6.14.0",
     "styled-components": "^6.0.0-rc.6"
   },
