+++ src/sections/Home/index.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material'
 const Home = () => {
-  return <Box>HELLO THERE</Box>
+  return <Box>HELLO THERE TESTED SUCCESSFULLY BY SAURABH</Box>
 }
 
 export default Home
